Spread out widgets and interfaces into separate files
