+++ src/main/interfaces/order-item.interface.ts
@@ -1,0 +1,6 @@
+import { Product } from "./product.interface";
+
+export interface OrderItem {
+    count: number;
+    product: Product;
+}
