+++ src/main/interfaces/order.interface.ts
@@ -1,0 +1,8 @@
+import { OrderItem } from "./order-item.interface";
+
+export interface Order {
+    items: OrderItem[];
+    count: number;
+    price: number;
+    timestamp?: string;
+}
