+++ src/main/interfaces/product.interface.ts
@@ -1,0 +1,8 @@
+export interface Product {
+    code: string;
+    title: string;
+    description: string;
+    price: number;
+    count: number;
+    sold: number;
+}
