+++ src/main/interfaces/user.interface.ts
@@ -1,0 +1,5 @@
+export interface User {
+    login: string;
+    name: number;
+    role: string;
+}
